[FIX] 전체적인 폴더구조 변경 - 사용자/관리자 고려 [UPDATE] web.xml DispathcerServlet 및 ContextLoadListener 구조 변경 [UPDATE] resources 관련 파일 받아들이는 경로 변경
