+++ src/main/java/com/bkc/main/controller/HomeController.java
@@ -1,4 +1,4 @@
-package com.bkc.controller;
+package com.bkc.main.controller;
 
 
 import org.springframework.stereotype.Controller;
@@ -10,6 +10,6 @@ public class HomeController {
 	
 	@RequestMapping(value = "/", method = RequestMethod.GET)
 	public String home() {
-		return "main/index";
+		return "index";
 	}
 }
