+++ src/main/webapp/WEB-INF/spring/deliveryServlet/delivery-servlet-context.xml
@@ -17,9 +17,9 @@
 
 	<!-- Resolves views selected for rendering by @Controllers to .jsp resources in the /WEB-INF/views directory -->
 	<beans:bean class="org.springframework.web.servlet.view.InternalResourceViewResolver">
-		<beans:property name="prefix" value="/WEB-INF/views/" />
+		<beans:property name="prefix" value="/WEB-INF/views/delivery/" />
 		<beans:property name="suffix" value=".jsp" />
 	</beans:bean>
 	
-	<context:component-scan base-package="com.bkc.controller" />
+	<context:component-scan base-package="com.bkc.delivery" />
 </beans:beans>
