+++ src/main/webapp/WEB-INF/web.xml
@@ -2,39 +2,81 @@
 <web-app xmlns="http://xmlns.jcp.org/xml/ns/javaee" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
 	xsi:schemaLocation="http://xmlns.jcp.org/xml/ns/javaee http://xmlns.jcp.org/xml/ns/javaee/web-app_3_1.xsd"
 	version="3.1">
+	
 	<display-name>bkc</display-name>
 	<absolute-ordering />
 	
+	<!-- 공유되는 서블릿 -->
 	<context-param>
 		<param-name>contextConfigLocation</param-name>
 		<param-value>
 			/WEB-INF/spring/appServlet/dao-context.xml
 			/WEB-INF/spring/appServlet/service-context.xml
-			/WEB-INF/spring/appServlet/servlet-context.xml
+			/WEB-INF/spring/appServlet/security-context.xml
 		</param-value>
 	</context-param>
-	
 	<listener>
 		<listener-class>org.springframework.web.context.ContextLoaderListener
 		</listener-class>
 	</listener>
+	
+	<!-- 메인 서블릿  -->
 	<servlet>
-		<servlet-name>appServlet</servlet-name>
+		<servlet-name>mainServlet</servlet-name>
 		<servlet-class>org.springframework.web.servlet.DispatcherServlet
 		</servlet-class>
 		<init-param>
 			<param-name>contextConfigLocation</param-name>
-			<param-value>/WEB-INF/spring/appServlet/servlet-context.xml
+			<param-value>
+				/WEB-INF/spring/appServlet/main-servlet-context.xml
 			</param-value>
 		</init-param>
 		<load-on-startup>1</load-on-startup>
 	</servlet>
-
 	<servlet-mapping>
-		<servlet-name>appServlet</servlet-name>
+		<servlet-name>mainServlet</servlet-name>
 		<url-pattern>/</url-pattern>
 	</servlet-mapping>
+	
+	<!-- 사용자 서블릿  -->
+	<servlet>
+		<servlet-name>clientServlet</servlet-name>
+		<servlet-class>org.springframework.web.servlet.DispatcherServlet
+		</servlet-class>
+		<init-param>
+			<param-name>contextConfigLocation</param-name>
+			<param-value>
+				/WEB-INF/spring/subpagesServlet/subpages-servlet-context.xml
+				/WEB-INF/spring/deliveryServlet/delivery-servlet-context.xml
+			</param-value>
+		</init-param>
+		<load-on-startup>2</load-on-startup>
+	</servlet>
+	<servlet-mapping>
+		<servlet-name>clientServlet</servlet-name>
+		<url-pattern>*.do</url-pattern>
+	</servlet-mapping>
+	
+	<!-- 관리자  -->
+	<servlet>
+		<servlet-name>adminServlet</servlet-name>
+		<servlet-class>org.springframework.web.servlet.DispatcherServlet
+		</servlet-class>
+		<init-param>
+			<param-name>contextConfigLocation</param-name>
+			<param-value>
+				/WEB-INF/spring/adminServlet/admin-servlet-context.xml
+			</param-value>
+		</init-param>
+		<load-on-startup>2</load-on-startup>
+	</servlet>
 
+	<servlet-mapping>
+		<servlet-name>adminServlet</servlet-name>
+		<url-pattern>*.mdo</url-pattern>
+	</servlet-mapping>
+
+	
 	<!-- EncodingFilter -->
 	<filter>
 		<filter-name>encodingFilter</filter-name>
@@ -49,6 +91,4 @@
 		<filter-name>encodingFilter</filter-name>
 		<url-pattern>/*</url-pattern>
 	</filter-mapping>
-
-
 </web-app>
