+++ src/main/webapp/resources/css/subpages/brand/newsNewItemview.css
