--- src/main/webapp/resources/js/menu/menu.css
@@ -1,211 +1,0 @@
-* {
-    list-style: none;
-}
-a, div, li{
-    box-sizing: border-box;
-}
-.burger{
-    padding: 50px 0 45px;
-}
-.main{
-    max-width: 1184px;
-    margin: 0 auto;
-    padding: 0 20px;
-}
-.menu{
-    position: relative;
-    margin-bottom: 26px;
-} 
-.menu .menu_burger{
-    position: absolute;
-    right: 0;
-    top: 50%;
-    margin: -17.5px 0 0;
-}
-.menu .menu_burger button{
-    font-size: 16px;
-    color: #b8b8b8;
-    display: inline-block;
-    font-weight: 700;
-    border: 0 solid transparent;
-    height: 35px;
-    border-width: 3px 0;
-    padding: 5px 0;
-}
-.menu .menu_preview{
-    margin-bottom: 15px;
-    font-size: 2.5rem;
-    line-height: 100%;
-    font-weight: 900;
-}
-.menu_burger{
-    overflow: auto;
-}
-.menu_burger li{
-    position: relative;
-    display: table-cell;
-    padding: 0 18px;
-    vertical-align: top;
-    white-space: nowrap;
-}
-.menu_burger li.on a, .menu_burger li.on button{
-    color: red;
-    border-bottom-color: red;
-}
-.menu_burger ul{
-    display: table;
-}
-button{
-    margin: 0;
-    vertical-align: middle;
-    overflow: visible;
-    background: 0;
-    appearance: auto;
-    text-rendering: auto;
-    border: 0;
-    outline: 0;
-    cursor: pointer;
-}
-.item{
-    margin-top: 60px;
-    padding-bottom: 46px;
-    overflow: hidden;
-}
-.item li{
-    position: relative;
-    float: left;
-    width: 25%;
-    margin-bottom: 8px;
-}
-.item .prd_img{
-    max-width: 220px;
-    height: 144px;
-    margin: 0 auto;
-    overflow: hidden;
-}
-.item .prd_img img{
-    width: 100%;
-}
-.item .cont{
-    min-height: 115px;
-    padding: 30px 5px 0;
-    text-align: center;
-}
-.item .tit {
-    font-weight: 700;
-    font-size: 1.25rem;
-}
-.btn_detail {
-    position: absolute;
-    left: 0;
-    top: 0;
-    width: 100%;
-    height: 100%;
-    text-indent: -99999em;
-    background: transparent;
-}
-a {
-    margin: 0;
-    color: #000;
-    text-decoration: none;
-    outline: none;
-    cursor: pointer;
-}
-@media (max-height: 500px) and (orientation: landscape), screen and (max-width: 768px) and (orientation: portrait){
-    .burger{
-        padding: 0!important;
-    }
-    .main{
-        min-width: 320px;
-        padding: 0;
-    }
-    .menu{
-        min-height: 56px;
-        margin: 0;
-        padding: 20px 0 10px;
-        background: #fff;
-    }
-    .menu .menu_preview{
-        margin: 0;
-        padding: 0 16px 10px;
-        font-size: 1.5rem;
-        text-align: center;
-        box-sizing: border-box;
-    }
-    .menu .menu_burger:last-child{
-        margin-bottom: -10px;
-    }
-    .menu .menu_burger{
-        position: static;
-        margin: 0;
-    }
-    .menu_burger li{
-        position: relative;
-        float: left;
-        width: 25%;
-        margin-bottom: -1px;
-    }
-    .menu_burger li, .menu_burger ul{
-        display: block;
-        border-bottom: 1px solid #f1f3f5;
-        text-align: center;
-    }
-    .menu_burger ul{
-        overflow: hidden;
-        width: 100%;
-    }
-    .menu_burger button{
-        text-align: center;
-        height: 49px;
-        padding: 0 10px;
-        border-width: 2px 0;
-    }
-    .menu .menu_burger button{
-        height: 56px;
-        border-width: 3px 0;
-        font-size: 1rem;
-    }
-    .item{
-        margin: 0;
-        padding: 0;
-    }
-    .item li{
-        position: relative;
-        float: none; 
-        display: flex;
-        -webkit-box-align: center;
-        align-items: center;
-        width: auto;
-        min-height: 95px;
-        margin-bottom: 0;
-        padding: 10px 40px 10px 133px;
-        border-bottom: 1px solid #f8f3ed;
-    }
-    .item .prd_img{
-        position: absolute;
-        left: 10px;
-        top: 50%;
-        width: 116px;
-        height: 76px;
-        margin-top: -38px;
-    }
-    .item .cont{
-        width: 100%;
-        min-height: 1em;
-        padding: 18px 0;
-        color: #2e2e2e;
-        text-align: left;
-    }
-    .item .tit{
-        font-size: 1rem;
-    }
-    .btn_detail {
-        position: absolute;
-        left: 0;
-        top: 0;
-        width: 100%;
-        height: 100%;
-        text-indent: -99999em;
-        background: transparent;
-    }
-}
